+++ camera/viewport.cpp
@@ -17,17 +17,17 @@ Viewport::Viewport()
 
 void Viewport::setFOV(float rads)
 {
-	scale = 0.5*Inv(Tan(rads*0.5));
+	scale = float(0.5*Inv(Tan(rads*0.5)));
 }
 
 float Viewport::getFOV() const
 {
-    return Atan(1.0/(2.0*scale))*2;
+    return float(Atan(1.0/(2.0*scale))*2);
 }
 
 float Viewport::getHFOV() const
 {
-    return Atan(1.0/(2.0*scale)*float(h)/float(w))*2;
+    return float(Atan(1.0/(2.0*scale)*float(h)/float(w))*2);
 }
 
 void Viewport::setPerspective(bool persp)
